requirements.txt
Django==2.2
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
django-tinymce==2.8.0
